Fixies for stupid bugs
So many changes for this pernicious insidious language when you dont have proper tooling

Apparently lesson learned regex string checks, the pattern HAS to be single quoted. The string content may or may not need to be double quote (hands in the air for random bug)

Also because you cant exit the sleep script from within a called subroutine if made more pure functional approach with return values to the iff function. I hate this debugging this language