Changed split to space - no difference
Makes no difference but I change the split to a space
The issue is that due to the pointless complexities of this scripting capability when using alias to call the function for the NON-CIDR variety you would have to pass both the IP and Netmask as separate parameters. 

Now do we want to add more confusing logic or restrict the dialog to only accept CIDR notations?